requirements.txt
pillow
pycountry
